Fix bugs related to existence of files and folders

+++ deploy.sh
@@ -3,7 +3,15 @@
 set -e
 
 for f in `find . -type f -not -path "./.git/*" -not -path "./deploy.sh" -not -path "./.gitignore" | cut -c3-`; do
-	rm ~/"$f"
+	if [ -f ~/"$f" ]; then
+		rm ~/"$f"
+	fi
+
+	# If dir exists in dotfiles but not in ~ => create it
+	if [ -d ${f%/*} ] && [ ! -d ~/${f%/*} ]; then
+		mkdir ~/${f%/*}
+	fi
+
 	cp "$f" ~/"$f"
 done
